+++ client/src/components/picture/AddOrCreateCollection.js
@@ -5,39 +5,19 @@ import AddToNew from './AddToNew'
 
 const AddOrCreateCollection = (props) => {
   const [ toggle, setToggle ] = useState(true)
-
   const addToExisting = () => setToggle(true)
   const addToNew = () => setToggle(false)
 
   return (
     <>
       { toggle 
-      
-      ?      <button onClick={addToNew}> Add To New Collection </button>
-
-
-      :      <button onClick={addToExisting}> Add To Existing Collection </button>
-
-      
+        ? <button onClick={addToNew}> Add To New Collection </button>
+        : <button onClick={addToExisting}> Add To Existing Collection </button>
       }
-      
       { toggle 
-      
-      
-      ? <AddToExisting toggleModal={props.toggleModal}/>
-      
-      
-      
-      : <AddToNew toggleModal={props.toggleModal}/>
-      
-      
+        ? <AddToExisting toggleModal={props.toggleModal}/>
+        : <AddToNew toggleModal={props.toggleModal}/>
       }
-      
-        
-      
-      
-
-      
     </>
   )
 }
