+++ client/src/components/picture/AddToNew.js
@@ -1,15 +1,12 @@
 import React, { useState } from 'react'
 import CollectionForm from '../collection/CollectionForm'
-import { CollectionConsumer } from '../../providers/CollectionProvider'
 import { ImageConsumer } from '../../providers/ImageProvider'
 
 
 const AddToNew = (props) => {
   const handleRes = (collection) => {
-    // props.createCollectionPicture(props.imageId, collection.id)
     props.addImageToCollection(collection.id)
     props.toggleModal()
-    
   }
 
   return(
@@ -19,15 +16,10 @@ const AddToNew = (props) => {
   )
 }
 
-const ConnectedAddToNew = (props) => ( //Dont need collection consumer
-  <CollectionConsumer>
-      {(value) => <AddToNew {...props} {...value} />}
-  </CollectionConsumer>
-)
-const ImageConnectedAddToExisting = (props) => (
+const ConnectedAddToNew = (props) => ( 
   <ImageConsumer>
-    {(value) => <ConnectedAddToNew {...props} {...value} />}
+      {(value) => <AddToNew {...props} {...value} />}
   </ImageConsumer>
-);
+)
 
-export default ImageConnectedAddToExisting
+export default ConnectedAddToNew
