+++ client/src/index.js
@@ -8,22 +8,19 @@ import { ImageProvider } from './providers/ImageProvider'
 
 import 'semantic-ui-css/semantic.min.css';
 import { initMiddleware } from 'devise-axios';
-import { CollectionProvider } from './providers/CollectionProvider';
 
 
 initMiddleware()
 ReactDOM.render(
   <React.StrictMode>
    <AuthProvider>
-    <CollectionProvider>
-      <FeedProvider>
-          <ImageProvider>
-            <BrowserRouter>
-              <App />
-            </BrowserRouter>
-        </ImageProvider>
-      </FeedProvider>
-    </CollectionProvider>
+    <FeedProvider>
+        <ImageProvider>
+          <BrowserRouter>
+            <App />
+          </BrowserRouter>
+      </ImageProvider>
+    </FeedProvider>
   </AuthProvider>
   </React.StrictMode>,
   document.getElementById('root')
